requirements.txt
Pillow>=9.0.0
pandas>=1.3.0
requests>=2.25.0
emoji>=2.0.0
openpyxl>=3.0.0